fix deleteProduct error handlers

+++ backend/controllers/product.controller.js
@@ -49,12 +49,16 @@ export const updateProduct = async (req, res) =>{
 
 export const deleteProduct = async (req, res) => {
     const{id} = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: "Invalid Product Id" }); // code 404 for not found
+    }
     
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success:true, message: "Product deleted"}); // code 200 for new deletion
     } catch (error) {
         console.log("error in deleting product",error.message);
-        res.status(404).json({ success:true, message: "Product not found"}); 
+        res.status(500).json({ success:true, message: "Server Error"}); 
     }
 };
